+++ lib/src/values/typedefs.dart
@@ -35,3 +35,5 @@ typedef VoidCallBackWithFuture = Future<void> Function();
 typedef StringsCallBack = void Function(String emoji, String messageId);
 typedef StringWithReturnWidget = Widget Function(String separator);
 typedef DragUpdateDetailsCallback = void Function(DragUpdateDetails);
+typedef ReactionCallback = void Function(Message message, String emoji);
+