+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [1.3.2] (Unreleased)
+
+* **Fix**: [165](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/issues/165) Fix issue
+  of user reaction callback provides incorrect message object when user react on any message 
+  with double or from reaction sheet.
 * **Fix**: [126](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/issues/126) Added
   flag to hide user name in chat.
 * **Feat**: [161](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/161) Added